change day of lateness fro int to long to be coherent with the rest of dates

+++ src/main/java/fr/d2factory/libraryapp/member/Member.java
@@ -46,7 +46,7 @@ public abstract class Member {
     /**
      * The number of day at which point the member becomes late
      */
-    public abstract int dayOfLateness();
+    public abstract long dayOfLateness();
 
     public BigDecimal getWallet() {
         return wallet;
